Refactor: extract shared post-training hooks and update SFT implementation
